"Map: Added high- and low NoisePollution, api fix"
