--- polygons.sql
@@ -1,35 +1,0 @@
-INSERT INTO polygons (id, name) VALUES (1, 'Polygon 1');
-INSERT INTO polygons (id, name) VALUES (2, 'Polygon 2');
-INSERT INTO polygons (id, name) VALUES (3, 'Polygon 3');
-INSERT INTO polygons (id, name) VALUES (4, 'Polygon 4');
-INSERT INTO polygons (id, name) VALUES (5, 'Polygon 5');
-INSERT INTO polygons (id, name) VALUES (6, 'Polygon 6');
-INSERT INTO polygons (id, name) VALUES (7, 'Polygon 7');
-INSERT INTO polygons (id, name) VALUES (8, 'Polygon 8');
-INSERT INTO polygons (id, name) VALUES (9, 'Polygon 9');
-INSERT INTO polygons (id, name) VALUES (10, 'Polygon 10');
-INSERT INTO polygons (id, name) VALUES (11, 'Polygon 11');
-INSERT INTO polygons (id, name) VALUES (12, 'Polygon 12');
-INSERT INTO polygons (id, name) VALUES (13, 'Polygon 13');
-INSERT INTO polygons (id, name) VALUES (14, 'Polygon 14');
-INSERT INTO polygons (id, name) VALUES (15, 'Polygon 15');
-INSERT INTO polygons (id, name) VALUES (16, 'Polygon 16');
-INSERT INTO polygons (id, name) VALUES (17, 'Polygon 17');
-INSERT INTO polygons (id, name) VALUES (18, 'Polygon 18');
-INSERT INTO polygons (id, name) VALUES (19, 'Polygon 19');
-INSERT INTO polygons (id, name) VALUES (20, 'Polygon 20');
-INSERT INTO polygons (id, name) VALUES (21, 'Polygon 21');
-INSERT INTO polygons (id, name) VALUES (22, 'Polygon 22');
-INSERT INTO polygons (id, name) VALUES (23, 'Polygon 23');
-INSERT INTO polygons (id, name) VALUES (24, 'Polygon 24');
-INSERT INTO polygons (id, name) VALUES (25, 'Polygon 25');
-INSERT INTO polygons (id, name) VALUES (26, 'Polygon 26');
-INSERT INTO polygons (id, name) VALUES (27, 'Polygon 27');
-INSERT INTO polygons (id, name) VALUES (28, 'Polygon 28');
-INSERT INTO polygons (id, name) VALUES (29, 'Polygon 29');
-INSERT INTO polygons (id, name) VALUES (30, 'Polygon 30');
-INSERT INTO polygons (id, name) VALUES (31, 'Polygon 31');
-INSERT INTO polygons (id, name) VALUES (32, 'Polygon 32');
-INSERT INTO polygons (id, name) VALUES (33, 'Polygon 33');
-INSERT INTO polygons (id, name) VALUES (34, 'Polygon 34');
-INSERT INTO polygons (id, name) VALUES (35, 'Polygon 35');
